refactor  listener structure
