--- src/kvstore/plugins/CMakeLists.txt
@@ -1,5 +1,0 @@
-# Copyright (c) 2022 vesoft inc. All rights reserved.
-#
-# This source code is licensed under Apache 2.0 License.
-
-nebula_add_subdirectory(elasticsearch)
