+++ src/storage/test/CMakeLists.txt
@@ -18,6 +18,7 @@ set(storage_test_deps
     $<TARGET_OBJECTS:internal_storage_client_obj>
     $<TARGET_OBJECTS:kvstore_obj>
     $<TARGET_OBJECTS:raftex_obj>
+    $<TARGET_OBJECTS:listener_obj>
     $<TARGET_OBJECTS:wal_obj>
     $<TARGET_OBJECTS:disk_man_obj>
     $<TARGET_OBJECTS:keyutils_obj>
